# ProjectsCompetitionPlatform/project-competition-openapi-specification.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi specification for \"Project competitions platform\" ",
    "description": " OpenAPI documentation for Spring Boot project called \"Project competitions platform\" that allows users to compete with each other and to win valuable prizes.",
    "contact": {
      "name": "Project-competition-platform team",
      "email": "project-competitions@mail.com"
    },
    "license": {
      "name": "Ukma license"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local ENV"
    }
  ],
  "paths": {
...
    "/api/images/{id}": {
      "get": {
        "tags": [
          "Images"
        ],
        "description": "Returns a specific image record by unique identifier in url.",
        "operationId": "findById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful request (image found)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Failed request (image not found)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageNotFoundException"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Images"
        ],
        "description": "Updates a specific image record by unique identifier in url.",
        "operationId": "updateById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageUpdateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully updated (image exists)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Failed request (image not found)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageNotFoundException"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Images"
        ],
        "description": "Updates a specific image record by unique identifier in url.",
        "operationId": "deleteById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully updated (image exists)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Failed request (image not found)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageNotFoundException"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/competitions/{id}": {
      "get": {
        "tags": [
          "competition-controller"
        ],
        "operationId": "findById_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CompetitionDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "competition-controller"
        ],
        "operationId": "updateById_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompetitionDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CompetitionDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "competition-controller"
        ],
        "operationId": "deleteById_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/api/images/_upload": {
      "post": {
        "tags": [
          "Images"
        ],
        "description": "Uploads image from form-data as a request body to the cloud storage \"Cloudinary\"",
        "operationId": "save",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully uploaded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request, when a provided file is null",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationFailDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ImageResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "url": {
            "type": "string"
          },
          "publicId": {
            "type": "string"
          }
        }
      },
      "ImageNotFoundException": {
        "type": "object",
        "properties": {
          "cause": {
            "type": "object",
            "properties": {
              "stackTrace": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "classLoaderName": {
                      "type": "string"
                    },
                    "moduleName": {
                      "type": "string"
                    },
                    "moduleVersion": {
                      "type": "string"
                    },
                    "methodName": {
                      "type": "string"
                    },
                    "fileName": {
                      "type": "string"
                    },
                    "lineNumber": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "className": {
                      "type": "string"
                    },
                    "nativeMethod": {
                      "type": "boolean"
                    }
                  }
                }
              },
              "message": {
                "type": "string"
              },
              "suppressed": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "stackTrace": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "classLoaderName": {
                            "type": "string"
                          },
                          "moduleName": {
                            "type": "string"
                          },
                          "moduleVersion": {
                            "type": "string"
                          },
                          "methodName": {
                            "type": "string"
                          },
                          "fileName": {
                            "type": "string"
                          },
                          "lineNumber": {
                            "type": "integer",
                            "format": "int32"
                          },
                          "className": {
                            "type": "string"
                          },
                          "nativeMethod": {
                            "type": "boolean"
                          }
                        }
                      }
                    },
                    "message": {
                      "type": "string"
                    },
                    "localizedMessage": {
                      "type": "string"
                    }
                  }
                }
              },
              "localizedMessage": {
                "type": "string"
              }
            }
          },
          "stackTrace": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "classLoaderName": {
                  "type": "string"
                },
                "moduleName": {
                  "type": "string"
                },
                "moduleVersion": {
                  "type": "string"
                },
                "methodName": {
                  "type": "string"
                },
                "fileName": {
                  "type": "string"
                },
                "lineNumber": {
                  "type": "integer",
                  "format": "int32"
                },
                "className": {
                  "type": "string"
                },
                "nativeMethod": {
                  "type": "boolean"
                }
              }
            }
          },
          "fileId": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "suppressed": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "stackTrace": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "classLoaderName": {
                        "type": "string"
                      },
                      "moduleName": {
                        "type": "string"
                      },
                      "moduleVersion": {
                        "type": "string"
                      },
                      "methodName": {
                        "type": "string"
                      },
                      "fileName": {
                        "type": "string"
                      },
                      "lineNumber": {
                        "type": "integer",
                        "format": "int32"
                      },
                      "className": {
                        "type": "string"
                      },
                      "nativeMethod": {
                        "type": "boolean"
                      }
                    }
                  }
                },
                "message": {
                  "type": "string"
                },
                "localizedMessage": {
                  "type": "string"
                }
              }
            }
          },
          "localizedMessage": {
            "type": "string"
          }
        }
      },
      "ImageUpdateDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          }
        }
      },
      "CompetitionDto": {
        "required": [
          "competition_begin_date",
          "competition_name",
          "has_prize_pool",
          "voting_begin_date",
          "voting_end_date"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "priceDescription": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          },
          "competition_name": {
            "maxLength": 100,
            "minLength": 3,
            "type": "string"
          },
          "competition_description": {
            "maxLength": 1000,
            "minLength": 0,
            "type": "string"
          },
          "competition_begin_date": {
            "type": "string",
            "format": "date-time"
          },
          "voting_begin_date": {
            "type": "string",
            "format": "date-time"
          },
          "voting_end_date": {
            "type": "string",
            "format": "date-time"
          },
          "has_prize_pool": {
            "type": "boolean"
          },
          "prize_pool": {
            "minimum": 0,
            "type": "number",
            "format": "double"
          },
          "competition_images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Image"
            }
          },
          "associated_projects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Project"
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tag"
            }
          },
          "payments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Payment"
            }
          }
        }
      },
      "CompetitionEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "beginDate": {
            "type": "string",
            "format": "date-time"
          },
          "votingBeginDate": {
            "type": "string",
            "format": "date-time"
          },
          "votingEndDate": {
            "type": "string",
            "format": "date-time"
          },
          "hasPrizePool": {
            "type": "boolean"
          },
          "priceDescription": {
            "type": "string"
          },
          "prizePool": {
            "type": "number",
            "format": "double"
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Image"
            }
          },
          "projects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Project"
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tag"
            }
          },
          "payments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Payment"
            }
          },
          "new": {
            "type": "boolean"
          }
        }
      },
      "Image": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "url": {
            "type": "string"
          },
          "publicId": {
            "type": "string"
          },
          "competitionEntities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompetitionEntity"
            }
          },
          "projects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Project"
            }
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Project"
            }
          },
          "new": {
            "type": "boolean"
          }
        }
      },
      "Payment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "sum": {
            "type": "number",
            "format": "double"
          },
          "paymentStatus": {
            "type": "string",
            "enum": [
              "SUCCESSFUL",
              "IN_PROCESS",
              "FAILED"
            ]
          },
          "failReason": {
            "type": "string"
          },
          "competition": {
            "$ref": "#/components/schemas/CompetitionEntity"
          },
          "user": {
            "$ref": "#/components/schemas/UserEntity"
          },
          "new": {
            "type": "boolean"
          }
        }
      },
      "Project": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserEntity"
          },
          "competitionEntities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompetitionEntity"
            }
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Image"
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tag"
            }
          },
          "new": {
            "type": "boolean"
          }
        }
      },
      "Tag": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "name": {
            "type": "string"
          },
          "projects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Project"
            }
          },
          "competitionEntities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompetitionEntity"
            }
          },
          "new": {
            "type": "boolean"
          }
        }
      },
      "UserEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "email": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "payments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Payment"
            }
          },
          "projects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Project"
            }
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Image"
            }
          },
          "new": {
            "type": "boolean"
          }
        }
      },
...
      "ValidationFailDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "exceptionClass": {
            "type": "string"
          },
          "exceptionTime": {
            "type": "string",
            "format": "date-time"
          },
          "fieldErrors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ImageRequestDto": {
        "required": [
          "image"
        ],
        "type": "object",
        "properties": {
          "image": {
            "type": "string",
            "format": "binary"
          }
        }
      }
    }
  }
}